LayoutTests/http/tests/media/video-seekable-stall-expected.txt

Test seekable range when video load is incomplete.

EVENT(canplay)
TEST(video.seekable) OK
TEST(video.seekable.length == 1) OK
TEST(video.seekable.start(0) == 0) OK
TEST(video.seekable.end(0) > 0) OK
TEST(video.seekable.end(0) < video.duration - 1) OK
TEST(video.currentTime = video.duration - 1) THROWS(DOMException.INDEX_SIZE_ERR) OK
END OF TEST
